Fixed several subtle bugs
Yaw ref now work due to change in spec, signal is actually active low. Combined with status check it seems to work fine now. Height had a bug where a uint was potentially subtracted from larger uint.  Flight controller now ignores changes to yaw at zero height. Over and out.

+++ height.c
@@ -80,7 +80,7 @@ void ZeroHeightTrigger(void) {
 
 int32_t GetHeight() {
     if (ref_found) {
-        return zero_reading - adc_val;
+        return (int32_t) zero_reading - adc_val;
     } else {
         return 0;
     }
